requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
charset-normalizer==2.1.1
commonmark==0.9.1
defusedxml==0.7.1
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
httpie==3.2.1
idna==3.4
multidict==6.0.2
psycopg2 @ file:///Users/builder/ci_310/psycopg2_1642535040511/work
Pygments==2.13.0
PySocks==1.7.1
pytz @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_1caf090c-01b6-4973-adac-d2eb884a674e7zonb9eb/croots/recipe/pytz_1654762641285/work
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.13
